feat(translator): migrate Netlify to Vercel
