--- netlify.toml
@@ -1,7 +1,0 @@
-[build]
-    functions = "apps/translator/functions"
-
-[[redirects]]
-    from = "/api/*"
-    to = "/.netlify/functions/:splat"
-    status = 200
